Bug in scriptLogsRotate.sh where the file to archive isn't exist

+++ tools/scriptLogsRotate.sh
@@ -8,5 +8,7 @@ HOMESNACK=/home/snack
 mv $HOMESNACK/logs/snacklog $HOMESNACK/logs/snacklog-$WEEKN
 service syslog-ng restart
 WEEKTOARCHIVE=$((WEEKN-4))
-cd $HOMESNACK/logs && tar jcvf snacklog-$WEEKTOARCHIVE.tar.bz2 snacklog-$WEEKTOARCHIVE
-rm $HOMESNACK/logs/snacklog-$WEEKTOARCHIVE
+if [ -f $HOMESNACK/logs/snacklog-$WEEKTOARCHIVE ]; then
+    cd $HOMESNACK/logs && tar jcvf snacklog-$WEEKTOARCHIVE.tar.bz2 snacklog-$WEEKTOARCHIVE
+    rm $HOMESNACK/logs/snacklog-$WEEKTOARCHIVE
+fi
